move colors up a directory
