+++ lib/reporters/mini.js
@@ -1,6 +1,6 @@
 'use strict';
 var logUpdate = require('log-update');
-var colors = require('./colors');
+var colors = require('../colors');
 var plur = require('plur');
 var beautifyStack = require('../beautify-stack');
