+++ lib/reporters/verbose.js
@@ -1,7 +1,7 @@
 'use strict';
 var prettyMs = require('pretty-ms');
 var figures = require('figures');
-var colors = require('./colors');
+var colors = require('../colors');
 var plur = require('plur');
 var beautifyStack = require('../beautify-stack');
